Individual Groundtruth Changes. One bug remains.
- Each algorithm now gets its own groundtruth TFR that is matched in aspect ratio but not size.

- Size of groundtruth TFR's is set so that their frequency dimension is => fmax-fmin/f_resolution, and their time dimension is => FS.

- The algorithmic TFRs are then rescaled up to size to match their groundtruth for statistical analysis.

- Everything works now, but in fixing things, I have realised that groundtruth will never be correct for CWT because CWT has a log scaled frequency vector, and groundtruth builds everything on linear vectors. This should have been obvious.

NOTE: In the code checked in here, the final plotted TFR's all look correct including CWT - this is because the frequency vectors stretch the image back to linear scaling. This is all well and good for the plots, but the matrices being compared for error analysis are still in different frequency scalings for CWT.

- Need to construct groundtruth in log scaling for CWT (switch).

- This commit finalises the last round of changes adding individual groundtruths for each TFR. Code is cleaned up and everything works except this log scaling issue.

Also, moved a number of obsolete scripts and functions to obsolete folder to cleanup repo.